# specs/002-backend-auth/contracts/api-contracts.yaml
openapi: 3.0.3
info:
  title: Secure Task Management API with Authentication
  description: API for managing tasks with JWT-based authentication and user isolation
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/auth/signup:
    post:
      summary: Register a new user
      description: Creates a new user account with secure password hashing and returns a JWT token
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securePassword123"
              required:
                - email
                - password
      responses:
        '200':
          description: User registered successfully, JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type:
                        type: string
                        example: "bearer"
        '400':
          description: Bad request - Invalid email format, weak password, or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      summary: Authenticate user and return JWT token
      description: Authenticates user credentials and returns a JWT token for subsequent requests
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  example: "securePassword123"
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful, JWT token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      token_type:
                        type: string
                        example: "bearer"
        '401':
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      summary: Get authenticated user information
      description: Returns information about the currently authenticated user
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: "123e4567-e89b-12d3-a456-426614174000"
                      email:
                        type: string
                        format: email
                        example: "user@example.com"
                      created_at:
                        type: string
                        format: date-time
                        example: "2023-01-01T00:00:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks:
    get:
      summary: Get all tasks for the authenticated user
      description: Retrieve a list of tasks belonging to the authenticated user
      tags:
        - tasks
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter tasks by completion status
          required: false
          schema:
            type: string
            enum: [all, pending, completed]
            default: all
        - name: sort
          in: query
          description: Sort tasks by specified criteria
          required: false
          schema:
            type: string
            enum: [created, title, due_date]
            default: created
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new task for the authenticated user
      description: Creates a new task associated with the authenticated user
      tags:
        - tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}:
    put:
      summary: Update an existing task
      description: Updates the properties of an existing task
      tags:
        - tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete a task
      description: Deletes a task belonging to the authenticated user
      tags:
        - tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: 'null'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/toggle:
    patch:
      summary: Toggle task completion status
      description: Toggles the completion status of a task
      tags:
        - tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          description: The ID of the task to toggle
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found or does not belong to the authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TaskBase:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          example: "Need to buy milk, bread, and eggs"
      required:
        - title
    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          example: "Need to buy milk, bread, and eggs"
        completed:
          type: boolean
          nullable: true
          example: false
    TaskResponse:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        completed:
          type: boolean
          example: false
        user_id:
          type: string
          format: uuid
          example: "987e6543-e89b-12d3-a456-426614174999"
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T00:00:00Z"
      required:
        - id
        - completed
        - user_id
        - created_at
        - updated_at
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
      required:
        - success
        - error